Refactor coordinate generation to adhere to Aero Protocol
- Removed hardcoded .chunk() and .chunk('auto') calls in GOES, NCEP GRIB, and IOAPI readers.
- Modernized GOES projection inversion using cached pyproj.Transformer for better Dask performance.
- Ensured coordinate chunking dynamically matches dataset chunks to preserve laziness.
- Integrated Aero Protocol compliance tests into existing test suites (test_cmaq.py, test_ncep_grib_modern.py, test_goes_opt.py).
- Updated provenance tracking in refactored functions.
